Create class Board and class Code
These classes are separate and there is no inheritance between them. The
Board class is mostly created as a placeholder to try and put a 4-digit
code into.

The Code class, when initialized, pushes a random number from 1-6 into
an array to create one of 4 digits in the code. This is done with a
'4.times do' and then those elements are joined.

Currently the number will generate successfully but will not show up in
the Board class object.