README update and compare method
This commit updates the README with how I want my version of mastermind
to be.

Currently the code creates an object of Board class, creates the 4-digit
code with new_code, and takes a user input after a prompt. The user
input is turned into an array to compare with the class variable
@@four_digit_code which was created in new_code method. The compare
method was made to test comparison of the user input guess and the
4-digit code class variable with an if..else block.

Currently there aren't issues with comparing the two.