Clue method, 12 turns, updates to puts statements
This commit changes the compare method from the last commit to a clue
method. The clue method firstly, decrements @@guesses which starts at
12 to keep track of how many guesses the user has left. Using an
if..else block, if guesses is 0 then the game is automatically over and
the user loses, elsif the 4-digit code and the user's guess matches, a
winning message is put to the terminal.

The rest of the clue method is a for loop utilizing a range, 0-3, for
each element in the 4-digit array and user guess. Depending on a match,
incorrect number, or correct in wrong place an 'O', 'X', or 'E',
respectively.

The program works correctly; first by briefly explaining the
instructions, prompting the user to take a guess, providing a clue and
how many guesses remain unless there is a 100% match or no more turns.