User selects to be code breaker or maker
This commit lets the user choose to be the code breaker or maker with
the class method make_or_break. A play_game class method is also created
which contains the main game (user guesses, receives clue).

The original initialization method messages are moved out of the class
for easier tracking of what's happening.

Currently there is a scope issue due to the play_game method calling a
method on an object that exists outside of the class definition. If the
user selects to be the code breaker, the undefined local variable error
shows up after the first guess is input. Selecting the code maker option
or inputting an invalid option works normally.