Class method cpu_solve and re-name clue method
This commit adds the cpu_solve method which is called when the user
chooses to make the code. Once the user has input their 4-digit code, it
is turned into an array and used as the argument for the method.

Inside the method, the CPU's first guess is a randomly generated 4-digit
code (same block from the new_code method) that is assigned to a
variable. The user's code and the CPU's first guess arrays are compared
using the same code from the formerly named clue method (now called
solve).

Currently the issue is that instead of producing four clues, it creates
ten.