cpu_solve method and undefined variable error fix
This commit fixed a quick undefined local variable error on line 40.
This was due to not changing 'x' and 'y' to 'code' and 'guess'
respectively.

Within the cpu_solve method, an if..elsif block is created to handle the
clue after the CPU's first guess. Three arrays were initialized before
the for loop to use: 'bank', 'e-bank', and 'next_guess'.

  *'bank' is used as a way to keep track of numbers that CPU can erase
  if an 'X' was given. It will be set up in setting up the next_guess.

  *'e-bank' contains the numbers that received an 'E' clue. It will be
  used to retry the number in a different 'E' clue position.

  *'next_guess' is the array that will be made using all the previous
  clues, numbers, etc.

Currently there are a lot of print statements at the end of the method
to monitor what each variable has.