Next steps in cpu_solve method
This commit changes the variable, 'next_guess', to 'pre_guess' because
this variable is the clue that the next_guess will based on.

After the CPU's takes the first guess and receives the first clue,
'pre_guess', it makes a guess, 'next_guess' using an if..elsif block to
figure out the next numbers to try.

  * If the pre_guess element is an Integer type, it will push that
  number to the next_guess array and move to the next element.

  * If the element is an 'E', it will sample an element from the e_bank.

  * If the element is an 'X', it will first subtract the e_bank array
  from the bank array and make a random selection from the numbers that
  are left. This is because an 'E' clue would've been given if the
  number existed somewhere in the code but an 'X' was given instead so
  a correct number must exist in the difference of the arrays.

After running the code several times and reviewing it, I am thinking of
where the loop needs to start and end for it to run on its own. I'm also
wondering if using a for loop might be better to use instead of an
if..elsif block for handling the clue.