Next_guess and change map method to for loop
This commit address the if..elsif block within the previous map method
that makes the CPU's next_guess. There were cases when the next_guess
would only be 3 digits. The issue was in the 'if' portion of the block.
The '==' was being used so when the CPU correctly guessed a number, it
wasn't being pushed to the next_guess array. This is changed to check
the class of the object with 'is_a?' (should be Integer to pass on to
next_guess).

The if..elsif block was also changed to be a for loop to iterate through
the range 0..3 so that it ensure that each digit in the 4-digit code was
being focused on.