Clue issues within cpu_solve method
The CPU's solving method has issues with coming up with better guesses
after receiving clues. I suspect this might be due to the clues not
being refreshed after a guess.

The CPU was able to solve the code a few times - very rarely within a
couple tries and a few times in 9 tries or more. I've noticed that if a
correct number is guessed within the first guess, it will remain locked
in for every guess after, which is a good thing. However, if an 'E' clue
is given on the first guess, that/those number(s) will keep being used
and never lock in even after getting it correct on a previous turn.