Cut out blocks within cpu_solve method
This commit removes a couple blocks within the cpu_solve method. This is
mainly because I started the loop earlier on to do away with repeating
blocks that only added more variables.

The pre_guess variable was changed to just 'guess', and next_guess was
removed to avoid confusion. The main variables within the class method
now are:

  * guesses - this keeps track of the CPU's guesses and decrements when
  it makes a guess.

  * guess - an array that contains the CPU's guesses.

  * user_code - this is an argument for cpu_solve which the entire
  method is constantly referring back to to see if there are matches.

  * clue - this variable is an array that the X's, E's and numbers are
  pushed to. This variable and the guess variable ping-pong with each
  other to create the next upcoming guess.

  * bank and e_bank - these are used as the CPU's options to choose from
  when creating its next guess.

Currently the main structure of the method is as follows:

  CPU takes its first guess...
    > IF guess == user_code
      CPU wins
    > ELSE
      > begin UNTIL loop
        > IF guess == user_code
            CPU wins
        > ELSIF guess != user_code
            create clue and use bank / e_bank to make better guess
        > ELSIF guesses == 0
            CPU loses
      > end UNTIL loop
    > END

CPU's solve rate is much better with some exceptions. There are more
issues to tend to such as declaring a CPU win or CPU lose.